reuse utc_now
